Remove concept-form from search result
