--- src/containers/SearchPage/components/results/SearchConcept/ConceptForm.tsx
@@ -1,190 +1,0 @@
-/**
- * Copyright (c) 2020-present, NDLA.
- *
- * This source code is licensed under the GPLv3 license found in the
- * LICENSE file in the root directory of this source tree.
- *
- */
-
-import { useFormik } from 'formik';
-import isEqual from 'lodash/fp/isEqual';
-import { useEffect } from 'react';
-import { useTranslation } from 'react-i18next';
-import styled from '@emotion/styled';
-import { ButtonV2 } from '@ndla/button';
-import { colors } from '@ndla/core';
-import { Select } from '@ndla/forms';
-import { Node } from '@ndla/types-taxonomy';
-import { InputField } from './SearchStyles';
-import AsyncSearchTags from '../../../../../components/Dropdown/asyncDropdown/AsyncSearchTags';
-import { MultiSelectDropdown } from '../../../../../components/Dropdown/MultiSelectDropdown';
-import MultiButton from '../../../../../components/MultiButton';
-import { CONCEPT_ADMIN_SCOPE, END_CONTROL, PUBLISHED } from '../../../../../constants';
-import { ConceptStatusType } from '../../../../../interfaces';
-import { fetchSearchTags } from '../../../../../modules/concept/conceptApi';
-import { useLicenses } from '../../../../../modules/draft/draftQueries';
-import { getLicensesWithTranslations } from '../../../../../util/licenseHelpers';
-import { useSession } from '../../../../Session/SessionProvider';
-
-export interface InlineFormConcept {
-  title: string;
-  license: string;
-  subjects: Node[];
-  tags: string[];
-  newStatus?: ConceptStatusType;
-}
-interface ErrorsType {
-  title?: string;
-  subjects?: string;
-}
-
-interface Props {
-  initialValues: InlineFormConcept;
-  status: string;
-  language: string;
-  onSubmit: (c: InlineFormConcept) => void;
-  allSubjects: Node[];
-  cancel: () => void;
-}
-
-const StyledHelpMessage = styled.span`
-  display: block;
-  color: ${colors.support.red};
-  margin-bottom: 5px;
-`;
-
-const validate = (values: InlineFormConcept): ErrorsType => {
-  const errors: ErrorsType = {};
-  if (!values.title) {
-    errors.title = 'required';
-  }
-  if (values.subjects.length === 0) {
-    errors.subjects = 'required';
-  }
-  return errors;
-};
-
-const ConceptForm = ({ initialValues, status, language, onSubmit, allSubjects, cancel }: Props) => {
-  const { t } = useTranslation();
-  const { userPermissions } = useSession();
-  const canPublish = !!userPermissions?.includes(CONCEPT_ADMIN_SCOPE);
-  const { data: licenses } = useLicenses({ placeholderData: [] });
-  const licensesWithTranslations = getLicensesWithTranslations(licenses!, language, true);
-  const formik = useFormik<InlineFormConcept>({
-    initialValues,
-    validate,
-    onSubmit,
-  });
-  const { values, errors, handleChange, handleBlur, setValues } = formik;
-  useEffect(() => {
-    setValues(initialValues);
-  }, [initialValues]); // eslint-disable-line react-hooks/exhaustive-deps
-  const hasChanges = !isEqual(initialValues, values);
-
-  const getStatus = (value: string, status: string) => {
-    if (value === 'saveAndPublish') {
-      return PUBLISHED;
-    } else if (status === PUBLISHED) {
-      return END_CONTROL;
-    } else {
-      return undefined;
-    }
-  };
-
-  const handleClick = (value: string) => {
-    const newStatus = getStatus(value, status);
-    onSubmit({ ...values, newStatus });
-  };
-
-  const secondaryButtons = canPublish
-    ? [
-        {
-          label: t('form.saveAndPublish'),
-          value: 'saveAndPublish',
-        },
-      ]
-    : [];
-
-  return (
-    <form>
-      <InputField>
-        <label htmlFor="title">{t('form.name.title')}</label>
-        <input
-          id="title"
-          name="title"
-          type="text"
-          aria-label={t('form.name.title')}
-          value={values.title}
-          onChange={handleChange}
-          onBlur={handleBlur}
-        />
-        {errors.title ? (
-          <StyledHelpMessage>
-            {t('validation.isRequired', { label: t('form.name.title') })}
-          </StyledHelpMessage>
-        ) : null}
-      </InputField>
-      <InputField>
-        <label htmlFor="license">{t('form.name.license')}</label>
-        <Select
-          value={values.license}
-          onChange={(v: any) => {
-            const value = v.target.value;
-            setValues({
-              ...values,
-              license: value,
-            });
-          }}
-        >
-          {!values.license && <option>{t('form.license.choose')}</option>}
-          {licensesWithTranslations.map((license) => (
-            <option value={license.license} key={license.license}>
-              {license.title}
-            </option>
-          ))}
-        </Select>
-      </InputField>
-      <InputField>
-        <label htmlFor="subjects">{t('searchForm.tagType.subjects')}</label>
-        <MultiSelectDropdown
-          name="subjects"
-          labelField="name"
-          value={values.subjects}
-          onChange={handleChange}
-          minSearchLength={1}
-          initialData={allSubjects}
-        />
-        {errors.subjects ? (
-          <StyledHelpMessage>
-            {t('validation.isRequired', { label: t('searchForm.tagType.subjects') })}
-          </StyledHelpMessage>
-        ) : null}
-      </InputField>
-      <InputField>
-        <label htmlFor="tags">{t('form.categories.label')}</label>
-        <AsyncSearchTags
-          multiSelect
-          language={language}
-          initialTags={values.tags}
-          updateValue={(updatedValue: string[]) => {
-            setValues({ ...values, tags: updatedValue });
-          }}
-          fetchTags={fetchSearchTags}
-        />
-      </InputField>
-      <div className="buttons">
-        <ButtonV2 className="form-button secondary" onClick={cancel}>
-          {t('editorFooter.cancelLabel')}
-        </ButtonV2>
-        <MultiButton
-          disabled={!hasChanges || Object.keys(errors).length > 0}
-          onClick={(value) => handleClick(value)}
-          mainButton={{ value: 'save', label: t(`form.save`) }}
-          secondaryButtons={secondaryButtons}
-        />
-      </div>
-    </form>
-  );
-};
-
-export default ConceptForm;
