--- src/containers/SearchPage/components/results/SearchConcept/index.ts
@@ -1,11 +1,0 @@
-/**
- * Copyright (c) 2020-present, NDLA.
- *
- * This source code is licensed under the GPLv3 license found in the
- * LICENSE file in the root directory of this source tree.
- *
- */
-
-import SearchConcept from './SearchConcept';
-
-export default SearchConcept;
